you get \> and you get \> and you get \>
outside of object>, qud only wants closed \> .  dunno when this change happened, but couldn't even get game to load on recent beta.  have issues with charli feet.  wanna wear arm gear yet have the arm slot go into a foot and not hand.  or maybe make a hand but it's a foot.